backend/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==44.0.0
cytoolz==1.1.0
eth-hash==0.7.1
eth-keys==0.7.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
grpcio==1.76.0
grpcio-tools==1.76.0
hiero-sdk-python==0.1.8
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
parsimonious==0.10.0
protobuf==6.33.0
psycopg2-binary==2.9.11
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
pyjnius==1.7.0
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.3
rlp==4.1.0
setuptools==80.9.0
SQLAlchemy==2.0.44
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
